изменение details, отселение employee в отдельный контроллер
