+++ WebStore/Views/Employees/_Layout.cshtml
@@ -1,0 +1,7 @@
+﻿@{ 
+    Layout = "../Shared/_Layout";
+}
+
+<div class="container">
+    @RenderBody()
+</div>
