--- WebStore/Views/Home/Sec.cshtml
@@ -1,13 +1,0 @@
-﻿
-<!DOCTYPE html>
-
-<html>
-<head>
-    <meta name="viewport" content="width=device-width" />
-    <title>Home Page</title>
-</head>
-<body>
-    <p>Home page</p>
-    <a href="~/Home/Employees">Список сотрудников</a>
-</body>
-</html>
