+++ WebStore/Views/Shared/_Layout.cshtml
@@ -5,8 +5,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="description" content="">
     <meta name="author" content="">
-    <title>@ViewBag.Title Home | E-Shopper</title>
-    @*<title>@ViewData["Title"] Home | E-Shopper</title>*@
+    <title>@ViewBag.Title | E-Shopper</title>
+    @*<title>@ViewData["Title"] | E-Shopper</title>*@
 
     @*@await Html.PartialAsync("Partial/_Styles")*@
     <partial name="Partial/_Styles" />
